Disable serde-arbitrary-precision feature of rust-decimal
...which enabled the arbitrary-precision feature of serde-json, which broke
yagna. This is because gsb-api relies on json integers being serialized
as flexbuffer numbers, while arbitrary-precision makes them serialize as
serde-json-specific structures.

This cannot be sensibly worked around, because cargo assumes crate
features are strictly additive and one cannot depend on the same library
twice with the only difference being the set of enabled features.

rust-decimal Decimals were only serialized in informational messages
which do not affect the correctness of the driver, so this change does
not affect any financial computations.